+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.1 under development
 
-- No changes in this release.
+- Consolidate Bootstrap5 views into `yiirocks/voyti-views-bootstrap5` package.
 
 ## 1.0.0 August 20, 2026